if source file does not exist, skip it

+++ src/Shopsys/ShopBundle/Component/Image/MigrateImageDoctrineListener.php
@@ -104,6 +104,9 @@ class MigrateImageDoctrineListener
         );
 
         try {
+            if ($this->filesystem->has($sourceFilepath) === false) {
+                return;
+            }
             if ($this->filesystem->has($targetFilepath)) {
                 $this->filesystem->delete($targetFilepath);
             }
